requirement.txt
tensorflow==2.4.2
tensorflow_estimator==2.4.0
tensorflow-serving-api==2.4.0
pytest==6.2.4
pyyaml==5.4.1